feat: structure change
changed application structure
