+++ pyproject.toml
@@ -16,6 +16,7 @@ pymysql = "^1.1.0"
 colorlog = "^6.7.0"
 gunicorn = "^21.2.0"
 waitress = "^2.1.2"
+flask-limiter = "^3.5.0"
 
 
 [build-system]
